需要登录时显示完整的登录窗口。
